using float color values instead of 0-255

+++ AuroraGSI-RL/TeamNode.h
@@ -5,9 +5,9 @@ class TeamNode {
 public:
 	int Index = -1;
 	int Goals = -1;
-	int Red = -1;
-	int Green = -1;
-	int Blue = -1;
+	float Red = -1;
+	float Green = -1;
+	float Blue = -1;
 	std::string Name = "";
 
 	nlohmann::json GetJson();
